Cassandra/requirements.txt
Flask>=3.0,<4.0
astrapy>=2.0,<3.0
python-dotenv>=1.0,<2.0
